lots of updates

+++ src/js/components/Toolbar.js
@@ -1,12 +1,31 @@
 import React, { Component } from "react";
 import { connect } from "react-redux";
 import '../../css/Toolbar.css'
-import { resetState, addCard, deleteCard } from "../actions/index";
+import { resetState, addCard, deleteCard, addNote, updateDeckElements, editNote } from "../actions/index";
 import {stateToCards} from "./stateToCards";
 import {SaveList} from "./SaveList";
 
 const mapStateToProps = state => {
+  var currentNote={};
+  var currentCard = state.cards.filter(oneCard => oneCard.cardNumber  === state.currentCardNumber)[0];
+  if (typeof currentCard === 'undefined') {
+    var currentCard= {paragraph: '',
+                  audioPath: '',
+                 notes: []}
+  }
+  var currentNote = currentCard.notes.filter(oneNote => oneNote.noteNumber === state.currentNoteNumber)[0];
+  if (typeof currentNote === 'undefined') {
+    var currentNote= {emphasis: false,
+                  noteNumber: 0,
+                  closed: false}
+  }
   return {
+    currentNoteEmphasis: currentNote.emphasis,
+    currentNoteNumber: currentNote.noteNumber,
+    currentNoteClosed: currentNote.closed,
+    currentNote: currentNote,
+    currentCard: currentCard,
+    notes: currentCard.notes,
     currentState: state,
     currentCardNumber: state.currentCardNumber
     }
@@ -14,9 +33,12 @@ const mapStateToProps = state => {
 
 function mapDispatchToProps(dispatch) {
   return {
+    addNote: noteInfo => dispatch(addNote(noteInfo)),
+    editNote: notes => dispatch(editNote(notes)),
     resetState: nothing => dispatch(resetState(nothing)),
     addCard: nothing => dispatch(addCard(nothing)),
-    deleteCard: nothing => dispatch(deleteCard(nothing))
+    deleteCard: nothing => dispatch(deleteCard(nothing)),
+    updateDeckElements: newDeckElements => dispatch(updateDeckElements(newDeckElements)),
   };
 }
 
@@ -30,37 +52,77 @@ class ToolbarDisplay extends Component {
     this.addCard = this.addCard.bind(this);
     this.deleteCard = this.deleteCard.bind(this);
     this.exportState = this.exportState.bind(this);
+    this.addNote = this.addNote.bind(this);
+    this.updateNote = this.updateNote.bind(this);
+    this.handleClose = this.handleClose.bind(this);
   }
 
+  updateNote(noteNumber, updatedMap) {
+    let newNotes = this.props.notes.map(singleNote => {
+          if (singleNote.noteNumber === noteNumber) {
+              return Object.assign({}, singleNote, updatedMap )
+            }
+          else {
+            return singleNote;
+          }});
+   return newNotes
+  }
 
  addCard(event) {
    this.props.addCard();
  }
 
  deleteCard(cardNumber) {
-   this.props.deleteCard();
+   if (window.confirm('Are you sure you wish to delete this item?')) {
+     this.props.deleteCard();
+   }
  }
+
+
+// <div className='delete-button' onClick={() => { if (window.confirm('Are you sure you wish to delete this item?')) this.onCancel(item) } } />
+
 
   exportState() {
     return stateToCards(this.props.currentState);
   }
 
   resetState() {
+  if (window.confirm('Are you sure you want to delete ALL CARDS?')) {
     this.props.resetState(null);
+  }
+  }
+
+  addNote(event) {
+    this.props.addNote({currentNotes: this.props.notes,
+                        currentCard: this.props.currentCard});
+  }
+
+  handleClose() {
+    var newNotes=this.updateNote(this.props.currentNoteNumber, {closed: !this.props.currentNoteClosed});
+    this.props.editNote({notes: newNotes});
+  }
+
+  handleEmphasis() {
+    var newNotes=this.updateNote(this.props.currentNoteNumber, {emphasis: !this.props.currentNoteEmphasis});
+    this.props.editNote({notes: newNotes});
   }
 
 
-
   render() {
-    const cardString = this.exportState();
-    console.log("cardString");
-    console.log(cardString)
   return (
-    <div className='toolbar'>
-      <button onClick={this.addCard}>Add Card</button>
-      <button onClick={this.deleteCard}>Delete</button>
-      <button onClick={this.resetState}>Reset</button>
+    <div>
+      <div className='toolbar d-grid gap-2 col-2 mx-auto'>
+      <button className="btn btn-info btn-sm"
+              id="addNote"
+              value="Add Note"
+              onClick={this.addNote} >Add Note</button>
+      <button className="btn btn-info btn-sm" onClick={(event) => this.handleClose()}>Close</button>
+      <button className="btn btn-info btn-sm" onClick={(event) => this.handleEmphasis()}>Emphasis</button>
+      <button className="btn btn-primary btn-sm" onClick={this.addCard}>Add Card</button>
+      <button className="btn btn-danger btn-sm" onClick={this.deleteCard}>Delete Card</button>
+      <button className="btn btn-danger btn-sm" onClick={this.resetState}>Reset Deck</button>
       <SaveList list={this.exportState()}/>
+      </div>
     </div>
   )
 }
@@ -69,3 +131,11 @@ class ToolbarDisplay extends Component {
 const Toolbar = connect(mapStateToProps, mapDispatchToProps)(ToolbarDisplay);
 
 export default Toolbar;
+
+
+/*
+<div className="inputDiv"> <input type="checkbox"
+                id="closed"
+                checked={el.closed}
+                onChange={(event) => this.handleClose(el.noteNumber, el.closed)}></input>
+                */
